requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
psycopg2-binary>=2.9.9
psycopg_pool>=3.2.0
pymilvus>=2.4.0
langchain-milvus>=0.1.0
langchain-community>=0.3.0
fastembed>=0.3.6
e2b>=1.0.5
python-dotenv>=1.0.1
rich>=13.7.1
